// blood/blood-frontend/src/shared/api/generated/model/createBloodRequestRequestUrgency.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Blood Donation Management API
 * Enterprise-ready modular monolith API for blood donation workflows, geo search, and SEO metadata.
 * OpenAPI spec version: 1.0.0
 */

export type CreateBloodRequestRequestUrgency = typeof CreateBloodRequestRequestUrgency[keyof typeof CreateBloodRequestRequestUrgency];


export const CreateBloodRequestRequestUrgency = {
  ROUTINE: 'ROUTINE',
  URGENT: 'URGENT',
  EMERGENCY: 'EMERGENCY',
} as const;